Refactor: extract SurveyLinkController
Extracts the link methods out of SurveyController in order to have a
distinct controller to create, refresh and delete survey links.

Now SurveyController is focused on CRUD operations on Surveys as
well as launching and stopping survey runs.
